update selector logic
